fix: :bug: update .hpps and .cpps from the logger implementation. Add them to the makefile.
