+++ includes/Logger.hpp
@@ -1,23 +1,13 @@
 #pragma once
 
-# include <iostream>
-
-enum    LogLevel {
-    DEBUG,
-    INFO,
-    WARNING,
-    ERROR
-};
-
-typedef struct s_logEvent {
-    enum LogLevel   level;
-    std::string     message;
-}t_logEvent;
+# include "WebservHeader.hpp"
 
 class LogHandler {
     public:
-        LogHandler() = default;
-        virtual ~LogHandler() = default;
+        // LogHandler() = default;
+        // virtual ~LogHandler() = default;
+        LogHandler();
+        virtual ~LogHandler();
         virtual void handleDebug(t_logEvent event) = 0;
         virtual void handleError(t_logEvent event) = 0;
         virtual void handleInfo(t_logEvent event) = 0;
@@ -31,8 +21,10 @@ class Logger {
         LogHandler      *_handler;
 
         Logger(void); // Construtor privado para não conseguirmos instanciar um novo Logger
-        Logger(const Logger &src) = delete; // Construtor de cópia deletado. Previnindo nova instancia
-        Logger &operator=(const Logger &src) = delete; // Operador de cópia deletado. Previnindo cópia da instancia
+        // Logger(const Logger &src) = delete; // Construtor de cópia deletado. Previnindo nova instancia
+        // Logger &operator=(const Logger &src) = delete; // Operador de cópia deletado. Previnindo cópia da instancia
+        Logger(const Logger &src); // Construtor de cópia deletado. Previnindo nova instancia
+        Logger &operator=(const Logger &src); // Operador de cópia deletado. Previnindo cópia da instancia
 
         Logger(enum LogLevel level, LogHandler *handler);
     public:
