+++ source/FileLogHandler.cpp
@@ -1,16 +1,14 @@
-#include "FileLogHandler.hpp"
+#include "../includes/WebservHeader.hpp"
 
-// FileLogHandler(void);
-FileLogHandler::FileLogHandler(std::string filename): LogHandler(), _logFile(filename, std::ofstream::out | std::ofstream::app) {}
+FileLogHandler::FileLogHandler(std::string filename): LogHandler(), _logFile(filename.c_str(), std::ofstream::out | std::ofstream::app) {}
 
-// ~FileLogHandler(void);
 FileLogHandler::~FileLogHandler(void) {_logFile.close();}
 
 // virtual void handleDebug(t_logEvent event);
 void FileLogHandler::handleDebug(t_logEvent event) {
     std::time_t timestamp = std::time(NULL);
     std::string result = std::ctime(&timestamp);
-    int pos = result.find("\n");
+    long unsigned int pos = result.find("\n");
 
     if (pos == std::string::npos) {
         _logFile << "[DEBUG] " << event.message << std::endl;
@@ -26,7 +24,7 @@ void FileLogHandler::handleDebug(t_logEvent event) {
 void FileLogHandler::handleError(t_logEvent event) {
     std::time_t timestamp = std::time(NULL);
     std::string result = std::ctime(&timestamp);
-    int pos = result.find("\n");
+    long unsigned int pos = result.find("\n");
 
     if (pos == std::string::npos) {
         _logFile << "[ERROR] " << event.message << std::endl;
@@ -42,7 +40,7 @@ void FileLogHandler::handleError(t_logEvent event) {
 void FileLogHandler::handleInfo(t_logEvent event) {
     std::time_t timestamp = std::time(NULL);
     std::string result = std::ctime(&timestamp);
-    int pos = result.find("\n");
+    long unsigned int pos = result.find("\n");
 
     if (pos == std::string::npos) {
         _logFile << "[INFO] " << event.message << std::endl;
@@ -58,7 +56,7 @@ void FileLogHandler::handleInfo(t_logEvent event) {
 void FileLogHandler::handleWarning(t_logEvent event) {
     std::time_t timestamp = std::time(NULL);
     std::string result = std::ctime(&timestamp);
-    int pos = result.find("\n");
+    long unsigned int pos = result.find("\n");
 
     if (pos == std::string::npos) {
         _logFile << "[WARNING] " << event.message << std::endl;
